Move topological GridTools functions to a new file.
