+++ source/grid/grid_tools.inst.in
@@ -82,11 +82,6 @@ for (deal_II_dimension : DIMENSIONS; deal_II_space_dimension : SPACE_DIMENSIONS)
                                                    deal_II_space_dimension>
           &triangulation);
 
-      template std::map<unsigned int, Point<deal_II_space_dimension>>
-      extract_used_vertices(
-        const Triangulation<deal_II_dimension, deal_II_space_dimension> &mesh,
-        const Mapping<deal_II_dimension, deal_II_space_dimension> &mapping);
-
       template std::pair<
         typename Triangulation<deal_II_dimension,
                                deal_II_space_dimension>::active_cell_iterator,
@@ -263,39 +258,6 @@ for (deal_II_dimension : DIMENSIONS; deal_II_space_dimension : SPACE_DIMENSIONS)
 #if deal_II_dimension <= deal_II_space_dimension
     namespace GridTools
     \{
-      template std::tuple<std::vector<Point<deal_II_space_dimension>>,
-                          std::vector<CellData<deal_II_dimension>>,
-                          SubCellData>
-      get_coarse_mesh_description(
-        const Triangulation<deal_II_dimension, deal_II_space_dimension> &tria);
-
-      template void
-      delete_unused_vertices(std::vector<Point<deal_II_space_dimension>> &,
-                             std::vector<CellData<deal_II_dimension>> &,
-                             SubCellData &);
-
-      template void
-      delete_duplicated_vertices(std::vector<Point<deal_II_space_dimension>> &,
-                                 std::vector<CellData<deal_II_dimension>> &,
-                                 SubCellData &,
-                                 std::vector<unsigned int> &,
-                                 double);
-
-      template void
-      invert_all_negative_measure_cells(
-        const std::vector<Point<deal_II_space_dimension>> &,
-        std::vector<CellData<deal_II_dimension>> &);
-
-      template std::size_t
-      invert_cells_with_negative_measure(
-        const std::vector<Point<deal_II_space_dimension>> &,
-        std::vector<CellData<deal_II_dimension>> &);
-
-#  if deal_II_dimension == deal_II_space_dimension
-      template void
-      consistently_order_cells(std::vector<CellData<deal_II_dimension>> &);
-#  endif
-
       template void
       shift<deal_II_dimension>(
         const Tensor<1, deal_II_space_dimension> &,
@@ -333,25 +295,6 @@ for (deal_II_dimension : DIMENSIONS; deal_II_space_dimension : SPACE_DIMENSIONS)
         Triangulation<deal_II_dimension, deal_II_space_dimension> &,
         const bool,
         const unsigned int);
-
-      template void
-      get_face_connectivity_of_cells(
-        const Triangulation<deal_II_dimension, deal_II_space_dimension>
-                               &triangulation,
-        DynamicSparsityPattern &cell_connectivity);
-
-      template void
-      get_vertex_connectivity_of_cells(
-        const Triangulation<deal_II_dimension, deal_II_space_dimension>
-                               &triangulation,
-        DynamicSparsityPattern &cell_connectivity);
-
-      template void
-      get_vertex_connectivity_of_cells_on_level(
-        const Triangulation<deal_II_dimension, deal_II_space_dimension>
-                               &triangulation,
-        const unsigned int      level,
-        DynamicSparsityPattern &cell_connectivity);
 
       template void
       partition_triangulation(
@@ -409,17 +352,6 @@ for (deal_II_dimension : DIMENSIONS; deal_II_space_dimension : SPACE_DIMENSIONS)
       template std::vector<bool>
       get_locally_owned_vertices(
         const Triangulation<deal_II_dimension, deal_II_space_dimension> &);
-
-      template std::map<unsigned int, Point<deal_II_space_dimension>>
-      get_all_vertices_at_boundary(
-        const Triangulation<deal_II_dimension, deal_II_space_dimension> &tria);
-
-      template std::vector<
-        std::set<Triangulation<deal_II_dimension,
-                               deal_II_space_dimension>::active_cell_iterator>>
-      vertex_to_cell_map(
-        const Triangulation<deal_II_dimension, deal_II_space_dimension>
-          &triangulation);
 
       template std::vector<std::vector<Tensor<1, deal_II_space_dimension>>>
       vertex_to_cell_centers_directions(
