+++ source/grid/grid_tools_topology.inst.in
@@ -1,0 +1,122 @@
+// ---------------------------------------------------------------------
+//
+// Copyright (C) 2001 - 2023 by the deal.II authors
+//
+// This file is part of the deal.II library.
+//
+// The deal.II library is free software; you can use it, redistribute
+// it, and/or modify it under the terms of the GNU Lesser General
+// Public License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+// The full text of the license can be found in the file LICENSE.md at
+// the top level directory of deal.II.
+//
+// ---------------------------------------------------------------------
+
+for (deal_II_dimension : DIMENSIONS; deal_II_space_dimension : SPACE_DIMENSIONS)
+  {
+#if deal_II_dimension <= deal_II_space_dimension
+    namespace GridTools
+    \{
+      template std::tuple<std::vector<Point<deal_II_space_dimension>>,
+                          std::vector<CellData<deal_II_dimension>>,
+                          SubCellData>
+      get_coarse_mesh_description(
+        const Triangulation<deal_II_dimension, deal_II_space_dimension> &tria);
+
+      template void
+      delete_unused_vertices(std::vector<Point<deal_II_space_dimension>> &,
+                             std::vector<CellData<deal_II_dimension>> &,
+                             SubCellData &);
+
+      template void
+      delete_duplicated_vertices(std::vector<Point<deal_II_space_dimension>> &,
+                                 std::vector<CellData<deal_II_dimension>> &,
+                                 SubCellData &,
+                                 std::vector<unsigned int> &,
+                                 double);
+
+#  if deal_II_dimension == deal_II_space_dimension
+      template void
+      delete_duplicated_vertices(std::vector<Point<deal_II_space_dimension>> &,
+                                 const double);
+#  endif
+
+      template void
+      invert_all_negative_measure_cells(
+        const std::vector<Point<deal_II_space_dimension>> &,
+        std::vector<CellData<deal_II_dimension>> &);
+
+      template std::size_t
+      invert_cells_with_negative_measure(
+        const std::vector<Point<deal_II_space_dimension>> &,
+        std::vector<CellData<deal_II_dimension>> &);
+
+#  if deal_II_dimension == deal_II_space_dimension
+      template void
+      consistently_order_cells(std::vector<CellData<deal_II_dimension>> &);
+#  endif
+
+      template std::map<unsigned int, Point<deal_II_space_dimension>>
+      get_all_vertices_at_boundary(
+        const Triangulation<deal_II_dimension, deal_II_space_dimension> &tria);
+
+      template std::vector<
+        std::set<Triangulation<deal_II_dimension,
+                               deal_II_space_dimension>::active_cell_iterator>>
+      vertex_to_cell_map(
+        const Triangulation<deal_II_dimension, deal_II_space_dimension>
+          &triangulation);
+
+      template std::map<unsigned int, Point<deal_II_space_dimension>>
+      extract_used_vertices(
+        const Triangulation<deal_II_dimension, deal_II_space_dimension> &mesh,
+        const Mapping<deal_II_dimension, deal_II_space_dimension> &mapping);
+
+      template void
+      get_face_connectivity_of_cells(
+        const Triangulation<deal_II_dimension, deal_II_space_dimension>
+                               &triangulation,
+        DynamicSparsityPattern &cell_connectivity);
+
+      template void
+      get_vertex_connectivity_of_cells(
+        const Triangulation<deal_II_dimension, deal_II_space_dimension>
+                               &triangulation,
+        DynamicSparsityPattern &cell_connectivity);
+
+      template void
+      get_vertex_connectivity_of_cells_on_level(
+        const Triangulation<deal_II_dimension, deal_II_space_dimension>
+                               &triangulation,
+        const unsigned int      level,
+        DynamicSparsityPattern &cell_connectivity);
+
+    \}
+#endif
+  }
+
+for (deal_II_dimension : DIMENSIONS)
+  {
+    template void
+    GridTools::remove_hanging_nodes<deal_II_dimension, deal_II_dimension>(
+      Triangulation<deal_II_dimension> & tria, bool, unsigned int);
+
+    template void
+    GridTools::remove_anisotropy<deal_II_dimension, deal_II_dimension>(
+      Triangulation<deal_II_dimension> &, double, unsigned int);
+
+#if deal_II_dimension < 3
+    template void
+    GridTools::remove_hanging_nodes<deal_II_dimension, deal_II_dimension + 1>(
+      Triangulation<deal_II_dimension, deal_II_dimension + 1> & tria,
+      bool,
+      unsigned int);
+
+    template void
+    GridTools::remove_anisotropy<deal_II_dimension, deal_II_dimension + 1>(
+      Triangulation<deal_II_dimension, deal_II_dimension + 1> &,
+      double,
+      unsigned int);
+#endif
+  }
